remove invalid characters in version
on FreeBSD, the version contains non-digit characters, such as
"-RELEASE". the method to compare the version throws an exception, but
it is not caught by the caller. thus, the default provider is choosen.

i'm not sure which is better to fix in ohai (breaks backward
compatibility) or in provider_mapping.rb, though.